changed price and discount fields to use BigDecimal, updated dtos and added mappers and validations to said dtos.

+++ src/main/java/OrderManager/Domain/Model/Item.java
@@ -8,6 +8,7 @@ import org.hibernate.annotations.Type;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
@@ -17,14 +18,15 @@ import java.util.UUID;
 @AllArgsConstructor
 @Builder
 public class Item {
+    //TODO:Seperate Item types with different fields.
     @Id
     @Type(type = "uuid-char")
     private UUID id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String size;
-    private double discount;
+    private BigDecimal discount;
     private boolean isAvailable;
     private int quantity;
